refactor: extract templates screen to layouts/templates.py
